+++ apps/www/registry/UNSW/example/breadcrumb-dropdown.tsx
@@ -7,13 +7,13 @@ import {
   BreadcrumbList,
   BreadcrumbPage,
   BreadcrumbSeparator,
-} from "@/registry/default/ui/breadcrumb"
+} from "@/registry/UNSW/ui/breadcrumb"
 import {
   DropdownMenu,
   DropdownMenuContent,
   DropdownMenuItem,
   DropdownMenuTrigger,
-} from "@/registry/default/ui/dropdown-menu"
+} from "@/registry/UNSW/ui/dropdown-menu"
 
 export default function BreadcrumbWithDropdown() {
   return (
