+++ apps/www/registry/UNSW/example/breadcrumb-ellipsis.tsx
@@ -8,7 +8,7 @@ import {
   BreadcrumbList,
   BreadcrumbPage,
   BreadcrumbSeparator,
-} from "@/registry/default/ui/breadcrumb"
+} from "@/registry/UNSW/ui/breadcrumb"
 
 export default function BreadcrumbCollapsed() {
   return (
